* commit 'a2c92e3894b43120ee7aa628cd39cbbaa01c0c25': (56 commits)
  Add bin/pkgscrpts files to MANIFEST.
  dd bin/pkgscrpts directory with scripts to package deployed unix binaries into tarfiles.
  Added .gitignore and .gitattributes files with contents from the last git mirror or the hdf5 repository prior to the manual migration of the hdf5 svn repository to the git repository in bitbucket. Added entries to MANIFEST for the 2 new files.
  Add VFD name to test folder
  Merge pull request #4 in HDFFV/hdf5 from ~LRKNOX/hdf5_lrk:hdf5_1_8 to hdf5_1_8
  Removed the -p option from the command to run autogen.sh.  It interfered with running chkmanifest on a machine with the correct autotools versions but without access to /usr/hdf/bin/AUTOTOOLS.  The autogen.sh script says the option is for HDF group use only and is likely to be removed.  The same effect can be achieved by adding /usr/hdf/bin/AUTOTOOLS to the beginning of PATH.
  Windows returns -1 instaed of 255 as on linux for the test
  Translate svn commands to git commands for chkmanifest and bbrelease scripts with source in git repository.
  [svn-r30321] HDFFV-9974: Correct parsing of UD filter parameters. Verified by user, who reported issue.
  [svn-r30319] Fix for daily test failure on emu -m64 for the checkin #30310. Tested on emu (-m64 & -m32), platypus, mayll, moohan, osx1010test, quail, kituo, ostrich, kite.
  [svn-r30318] whitespace cleanup
  [svn-r30317] Whitespace cleanup
  [svn-r30316] Use dns name for CDash instead of ip address
  [svn-r30315] HDFFV-9972: Fix test reference and change default for H5Iget_name in JNI API.
  [svn-r30314] HDFFV-9971: Change the usage of a CMake variable for the build type to the recommended variable name. Updated documentation.
  [svn-r30313] HDFFV-9972: unsatisfied link error under debug on Windows. Added windows name suffix for debug to CMake cmd_arg for examples and test. Fix issues discovered under debug testing;    Create a version of H5Iget_name that correctly returns the name.    Rework PIN_JAVA_STRING macro and usage to eliminate possible memory leaks by using if-else instead of mid-routine return.    Update example to use new H5Iget_name API.
  [svn-r30312] HDFFV-9706: Commit changes as discussed in HDF5-379.
  [svn-r30311] Purpose: Fix bug HDFFR-9920 cont. trunk: Description:     Continued rearranging the classes to model the relationship of HDF5     objects more accurately.  The changes included:     - in the baseclass list of Attribute, changed "public IdComponent" to       "public H5Location", because location sometime can be specified with       attribute     - moved H5A wrappers in H5Location to H5Object because H5A functions       can't be called on attribute id     - removed the stubs Attribute::iterateAttrs and Attribute::renameAttr     - removed Attribute::getFileName and Attribute::flush, because           H5Location has them     - result of the modified partial class diagram, regarding Attribute
  [svn-r30310] Fix daily test failure for previous checkin #30308. Tested on platypus and mayll.
  [svn-r30309] Purpose: Fix bug HDFFR-9920 Description:     Rearranged the classes to model the relationship of HDF5 objects more     accurately.  The changes included:     - moved CommonFG's methods to Group     - removed CommonFG from Group's base class list     - removed CommonFG from H5File's base class list and changed       "public H5Location" to "public Group" in the base class list     - CommonFG became unused     - result of the modified partial class diagram:
  ...

+++ src/H5Gnode.c
@@ -715,7 +715,7 @@ H5G_node_insert(H5F_t *f, hid_t dxpl_id, haddr_t addr,
 		md_key->offset = ent.name_off;
 	} /* end if */
         else {
-	    idx -= H5F_SYM_LEAF_K(f);
+	    idx -= (int)H5F_SYM_LEAF_K(f);
 	    insert_into = snrt;
 	    if(idx == (int)H5F_SYM_LEAF_K(f)) {
 		rt_key->offset = ent.name_off;
